requirements.txt

fastapi==0.111.0
httpx==0.27.0
orjson==3.10.3
pip==24.0
pydantic==2.7.2
pydantic_core==2.18.3
prometheus-client
prometheus-fastapi-instrumentator
uvicorn==0.30.0
virtualenv==20.26.2
requests==2.31.0
python-dotenv==1.0.1
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib==1.7.4
bcrypt<4.1.0
python-jose
asyncpg
 
 


